[F5Address]::IsMatch; Fix #129
-Implemented [F5Address] Equals operator override, which allows -eq,
-contains, etc. comparisons of [F5Address] objects. (I did not end up
using this for the conditional route domain filtering.  I implemented it
properly so that if a -eq b -and b -eq c then a -eq c -and c -eq a.)
-Implemented [F5Address]::IsMatch instance and static methods with
ipaddress and conditional route domain filtering.
-Replaced [IPAddress]::Any with [F5Address]::Any to avoid redundant type
coercion
-Test-Node: Fix #129: Now supports -Address, -Name, or both; I had to
change it to reflect the changes above, so I decided to kill 2 birds
with 1 stone after all.
-Disable/Enable/Get/Remove/Test-Node parameters and logic aligned
-Reformatted PoshLTM.Types.cs: no hanging brackets