Update Get-PoolsForMember.ps1
Pool members may not have IP addresses assigned and use a value of "any6". Since this is not a discrete IP address, Get-PoolsForMember will fail to return the correct results when a pool member object is passed in. Since pool member names must be unique, I added -Name as a param and use that instead of Address when the pool member object is passed.